Switch to extension method
