+++ CommandLineParser/CommandLineParser`TOption.cs
@@ -452,7 +452,7 @@ namespace MatthiWare.CommandLine
             {
                 var attributes = propInfo.GetCustomAttributes(true);
 
-                var lambda = GetLambdaExpression(propInfo, out string key);
+                var lambda = propInfo.GetLambdaExpression(out string key);
 
                 var actions = new List<Action>(4);
                 bool ignoreSet = false;
@@ -498,19 +498,6 @@ namespace MatthiWare.CommandLine
 
                 foreach (var action in actions)
                     action();
-            }
-
-            LambdaExpression GetLambdaExpression(PropertyInfo propInfo, out string key)
-            {
-                var entityType = propInfo.DeclaringType;
-                var propType = propInfo.PropertyType;
-                var parameter = Expression.Parameter(entityType, entityType.FullName);
-                var property = Expression.Property(parameter, propInfo);
-                var funcType = typeof(Func<,>).MakeGenericType(entityType, propType);
-
-                key = $"{entityType.ToString()}.{propInfo.Name}";
-
-                return Expression.Lambda(funcType, property, parameter);
             }
         }
     }
